Connect AREF to +5V.
No sensible way to route AREF, so for now the compromise is to just route
it to +5V so the ADC can be *usable* when configured to use an external
voltage source (e.g. for testing after programming) while not being
particularly accurate (since this is going to be noisy.)

Given that the primary purpose of this board is programming and then
post-programming brief testing, with the chip ultimately ending up in
a dedicated PCB design, this seems like a reasonable compromise.